+++ Server/API.cs
@@ -2715,7 +2715,7 @@ namespace GTANetworkServer
             msg.Write((byte)PacketType.ChatData);
             msg.Write(data.Length);
             msg.Write(data);
-            player.NetConnection.SendMessage(msg, NetDeliveryMethod.ReliableOrdered, (int)ConnectionChannel.Chat);
+            Program.ServerInstance.Send(msg, player.NetConnection, NetDeliveryMethod.ReliableOrdered, (int)ConnectionChannel.Chat);
         }
 
         public void setPlayerWantedLevel(Client player, int wantedLevel)